+++ src/updates/components/UpdatesLayout/UpdatesLayout.tsx
@@ -2,10 +2,10 @@ import { useNavigate } from "@tanstack/react-router";
 import dayjs from "dayjs";
 import { useMemo } from "react";
 import { colours } from "src/colours/colours.constant";
-import { ListSection } from "src/common/ListSection/ListSection";
-import { TableOfContentsListItem } from "src/common/TableOfContentsListItem/TableOfContentsListItem";
 import { Calendar } from "src/common/components/Calendar/Calendar";
 import { EmptyState } from "src/common/components/EmptyState/EmptyState";
+import { ListSection } from "src/common/components/ListSection/ListSection";
+import { TableOfContentsListItem } from "src/common/components/TableOfContentsListItem/TableOfContentsListItem";
 import { TwoPaneLayout } from "src/common/components/TwoPaneLayout/TwoPaneLayout";
 import { UpdateEditor } from "src/updates/components/UpdateEditor/UpdateEditor";
 import { UpdatesSection } from "src/updates/components/UpdatesSection/UpdatesSection";
